Made it so it shows '(new changes)' on the footer instead of opening the changelog modal immediately for users that have not seen new changes
