+++ packages/frontend/app/(dashboard)/components/FooterBar.tsx
@@ -1,8 +1,33 @@
+import { API } from '@/app/api'
 import ChangeLogModal from '@/app/components/ChangeLogModal'
+import { useUserInfo } from '@/app/contexts/userContext'
+import { getErrorMessage } from '@/app/utils/generalUtils'
 import { useState } from 'react'
 
 const FooterBar: React.FC = () => {
   const [isChangelogOpen, setIsChangelogOpen] = useState(false)
+
+  const { userInfo, setUserInfo } = useUserInfo()
+
+  const IReadTheChangelog = () => {
+    setUserInfo({ ...userInfo, readChangeLog: true })
+
+    const readChangelog = async () => {
+      if (!userInfo?.readChangeLog) {
+        try {
+          await API.profile.readChangelog()
+        } catch (error) {
+          const errorMessage = getErrorMessage(error)
+          console.error(
+            'Failed to tell server that changelog was read:',
+            errorMessage,
+          )
+        }
+      }
+    }
+
+    readChangelog()
+  }
 
   return (
     // Hide footer on mobile since screen space is more valuable
@@ -11,14 +36,17 @@ const FooterBar: React.FC = () => {
       aria-hidden="true"
     >
       <div>
-        Version 2.0.0{' '}
         <a
-          className="ml-1 cursor-pointer text-blue-600 underline"
+          className="ml-1 cursor-pointer "
           onClick={() => {
             setIsChangelogOpen(true)
+            IReadTheChangelog()
           }}
         >
-          What&apos;s new?
+          Version 1.1.1{' '}
+          {userInfo.readChangeLog === false && (
+            <span className="text-green-500">(New Changes!)</span>
+          )}
         </a>
         <ChangeLogModal
           isOpen={isChangelogOpen}
